chore(team): split step reconcile helpers
