+++ src/team/manager.rs
@@ -46,6 +46,7 @@ mod step_continuity;
 mod step_lifecycle;
 mod step_materialization;
 mod step_queries;
+mod step_reconcile;
 mod support;
 mod task_catalog;
 mod task_updates;
